isolate return path and flask dep on separate script
